Fixed several issues in triggers

+++ include/mind-net-plugin-repetition/mindnet/plugins/repetition/triggers/RSessionBeforeCreateTrigger.hpp
@@ -20,8 +20,8 @@ namespace mindnet::plugins::repetition::triggers
             api::OperationResult& validation_result,
             api::OperationResult& action_result,
             const model::ModelDefinition def,
-            int user_id,
-            int id,
+            i64 user_id,
+            i64 id,
             entity_fields& fields,
             entity_fields& old_fields,
             const orm::QueryParams query_params
